新增 自动注册 功能

+++ src/main/java/com/example/competition/dao/mapper/UserMapper.java
@@ -1,11 +1,11 @@
 package com.example.competition.dao.mapper;
 
+import com.example.competition.dao.dto.UserRoleDTO;
 import com.example.competition.dao.model.User;
 import com.example.competition.dao.model.UserExample;
 import java.util.List;
 import java.util.Map;
 
-import com.example.competition.dao.dto.UserRoleDTO;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Component;
 
@@ -15,7 +15,7 @@ public interface UserMapper {
 
     int deleteByExample(UserExample example);
 
-    int deleteByPrimaryKey(Long userId);
+    int deleteByPrimaryKey(Integer userId);
 
     int insert(User record);
 
@@ -23,7 +23,7 @@ public interface UserMapper {
 
     List<User> selectByExample(UserExample example);
 
-    User selectByPrimaryKey(Long userId);
+    User selectByPrimaryKey(Integer userId);
 
     int updateByExampleSelective(@Param("record") User record, @Param("example") UserExample example);
 
@@ -32,7 +32,6 @@ public interface UserMapper {
     int updateByPrimaryKeySelective(User record);
 
     int updateByPrimaryKey(User record);
-
     /** 查询用户和角色相关信息 */
     UserRoleDTO selectUserAndRole(Map<String,Object> map);
 }
